+++ src/main/java/com/unstoppabledomains/resolution/naming/service/NamingServiceType.java
@@ -1,6 +1,7 @@
 package com.unstoppabledomains.resolution.naming.service;
 
 public enum NamingServiceType {
-    CNS,
-    ZNS
+    UNS,
+    ZNS,
+    ENS
 }