+++ src/test/java/com/unstoppabledomains/resolution/ResolutionTest.java
@@ -4,9 +4,11 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.unstoppabledomains.TestUtils;
+import com.unstoppabledomains.config.network.NetworkConfigLoader;
 import com.unstoppabledomains.config.network.model.Network;
 import com.unstoppabledomains.exceptions.ns.NSExceptionCode;
 import com.unstoppabledomains.exceptions.ns.NamingServiceException;
+import com.unstoppabledomains.resolution.TokenUriMetadata.TokenUriMetadataAttribute;
 import com.unstoppabledomains.resolution.contracts.DefaultProvider;
 import com.unstoppabledomains.resolution.contracts.interfaces.IProvider;
 import com.unstoppabledomains.resolution.dns.DnsRecord;
@@ -18,8 +20,9 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -32,13 +35,20 @@ public class ResolutionTest {
 
     @BeforeAll
     public static void init() {
-        resolution = new Resolution();
+        resolution = Resolution.builder()
+        .chainId(NamingServiceType.ZNS, Network.MAINNET)
+        .providerUrl(NamingServiceType.UNS, TestUtils.TESTING_UNS_PROVIDER_URL)
+        .contractAddress(NamingServiceType.UNS, NetworkConfigLoader.getContractAddress(Network.RINKEBY, "ProxyReader"))
+        .providerUrl(NamingServiceType.ENS, TestUtils.TESTING_ENS_PROVIDER_URL)
+        .build();
     }
 
     @Test
-    public void resolveRinkebyDomain() throws Exception {
+    public void resolveTestnetDomain() throws Exception {
         DomainResolution rinkebyResolution = Resolution.builder()
-            .providerUrl(NamingServiceType.CNS, "https://rinkeby.infura.io/v3/e0c0cb9d12c440a29379df066de587e6")
+            .providerUrl(NamingServiceType.ENS, "https://mainnet.infura.io/v3/e0c0cb9d12c440a29379df066de587e6")
+            .providerUrl(NamingServiceType.UNS, "https://rinkeby.infura.io/v3/e0c0cb9d12c440a29379df066de587e6")
+            .contractAddress(NamingServiceType.UNS, NetworkConfigLoader.getContractAddress(Network.RINKEBY, "ProxyReader"))
             .build();
         String ethAddress = rinkebyResolution.getAddress("udtestdev-creek.crypto", "eth");
         assertEquals("0x1C8b9B78e3085866521FE206fa4c1a67F49f153A", ethAddress);
@@ -47,55 +57,102 @@ public class ResolutionTest {
 
     @Test
     public void testDifferentNetworks() throws Exception {
-        DomainResolution customCnsNetwork = Resolution.builder()
-            .providerUrl(NamingServiceType.CNS, "https://rinkeby.infura.io/v3/e0c0cb9d12c440a29379df066de587e6")
+        DomainResolution customNetworks = Resolution.builder()
+            .providerUrl(NamingServiceType.UNS, "https://rinkeby.infura.io/v3/e0c0cb9d12c440a29379df066de587e6")
+            .chainId(NamingServiceType.ENS, Network.GOERLI)
             .chainId(NamingServiceType.ZNS, Network.KOVAN)
             .build();
-            
-        Network customCnsChainId = customCnsNetwork.getNetwork(NamingServiceType.CNS);
-        Network customZnsChainId = customCnsNetwork.getNetwork(NamingServiceType.ZNS);
-        assertEquals(Network.RINKEBY, customCnsChainId);
+
+        Network customUnsChainId = customNetworks.getNetwork(NamingServiceType.UNS);
+        Network customEnsChainId = customNetworks.getNetwork(NamingServiceType.ENS);
+        Network customZnsChainId = customNetworks.getNetwork(NamingServiceType.ZNS);
+        assertEquals(Network.RINKEBY, customUnsChainId);
+        assertEquals(Network.GOERLI, customEnsChainId);
         assertEquals(Network.KOVAN, customZnsChainId);
     }
 
     @Test
     public void testDefaultNetworks() throws Exception {
         DomainResolution defaultSettings = new Resolution();
-        Network defaultCnsChainId = defaultSettings.getNetwork(NamingServiceType.CNS);
+        Network defaultUnsChainId = defaultSettings.getNetwork(NamingServiceType.UNS);
+        Network defaultEnsChainId = defaultSettings.getNetwork(NamingServiceType.ENS);
         Network defaultZnsChainId = defaultSettings.getNetwork(NamingServiceType.ZNS);
-        assertEquals(Network.MAINNET, defaultCnsChainId);
+        assertEquals(Network.MAINNET, defaultUnsChainId);
+        assertEquals(Network.MAINNET, defaultEnsChainId);
         assertEquals(Network.MAINNET, defaultZnsChainId);
     }
 
     @Test
     public void shouldResolveFromResolutionCreatedByBuilder() throws Exception {
         DomainResolution resolutionFromBuilder = Resolution.builder()
-                .chainId(NamingServiceType.CNS, Network.ROPSTEN)
-                .providerUrl(NamingServiceType.CNS, TestUtils.TESTING_CNS_PROVIDER_URL)
-                .build();
+        .chainId(NamingServiceType.UNS, Network.RINKEBY)
+        .chainId(NamingServiceType.ZNS, Network.MAINNET)
+        .chainId(NamingServiceType.ENS, Network.ROPSTEN)
+        .providerUrl(NamingServiceType.UNS, TestUtils.TESTING_UNS_PROVIDER_URL)
+        .providerUrl(NamingServiceType.ENS, TestUtils.TESTING_ENS_PROVIDER_URL)
+        .contractAddress(NamingServiceType.UNS, NetworkConfigLoader.getContractAddress(Network.RINKEBY, "ProxyReader"))
+        .build();
 
-        assertEquals("0x8aad44321a86b170879d7a244c1e8d360c99dda8", resolutionFromBuilder.getOwner("brad.crypto"));
+        assertEquals("0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2", resolutionFromBuilder.getOwner("testing.crypto"));
+        assertEquals("0x842f373409191cff2988a6f19ab9f605308ee462", resolutionFromBuilder.getOwner("monkybrain.eth"));
         assertEquals("0xcea21f5a6afc11b3a4ef82e986d63b8b050b6910", resolutionFromBuilder.getOwner("johnnyjumper.zil"));
     }
 
     @Test
     public void shouldResolveFromResolutionCreatedByBuilderWithInfura() throws Exception {
         DomainResolution resolutionFromBuilderWithInfura = Resolution.builder()
-                .chainId(NamingServiceType.CNS, Network.ROPSTEN)
-                .infura(NamingServiceType.CNS, Network.MAINNET, TestUtils.TESTING_INFURA_CNS_PROJECT_ID)
-                .build();
+            .chainId(NamingServiceType.ENS, Network.ROPSTEN)
+            .chainId(NamingServiceType.ZNS, Network.MAINNET)
+            .infura(NamingServiceType.ENS, TestUtils.TESTING_INFURA_ENS_PROJECT_ID)
+            .infura(NamingServiceType.UNS, Network.RINKEBY, TestUtils.TESTING_INFURA_UNS_PROJECT_ID)
+            .contractAddress(NamingServiceType.UNS, NetworkConfigLoader.getContractAddress(Network.RINKEBY, "ProxyReader"))
+            .build();
 
-        assertEquals("0x8aad44321a86b170879d7a244c1e8d360c99dda8", resolutionFromBuilderWithInfura.getOwner("brad.crypto"));
+        assertEquals("0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2", resolutionFromBuilderWithInfura.getOwner("testing.crypto"));
+        assertEquals("0x5d069edc8cc1c559e4482bec199c13547455208", resolutionFromBuilderWithInfura.getOwner("monkybrain.eth"));
         assertEquals("0xcea21f5a6afc11b3a4ef82e986d63b8b050b6910", resolutionFromBuilderWithInfura.getOwner("johnnyjumper.zil"));
     }
 
     @Test
-    public void isSupported() {
-        boolean isValid = resolution.isSupported("brad.crypto");
+    public void isSupported() throws NamingServiceException{
+        boolean isValid = resolution.isSupported("example.test");
+        assertFalse(isValid);
+
+        isValid = resolution.isSupported("example.tqwdest");
+        assertFalse(isValid);
+
+        isValid = resolution.isSupported("example.qwdqwdq.wd.tqwdest");
+        assertFalse(isValid);
+
+        isValid = resolution.isSupported("udtestdev-my-new-tls.wallet");
         assertTrue(isValid);
 
-        isValid = resolution.isSupported("brad.unsupported");
-        assertFalse(isValid);
+        isValid = resolution.isSupported("example.crypto");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.coin");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.wallet");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.bitcoin");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.x");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.888");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.nft");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.dao");
+        assertTrue(isValid);
+
+        isValid = resolution.isSupported("example.blockchain");
+        assertTrue(isValid);
     }
 
     @Test
@@ -105,6 +162,11 @@ public class ResolutionTest {
         hash = resolution.getNamehash("brad.crypto");
         assertEquals("0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9", hash);
 
+        hash = resolution.getNamehash("wallet");
+        assertEquals("0x1e3f482b3363eb4710dae2cb2183128e272eafbe137f686851c1caea32502230", hash);
+        hash = resolution.getNamehash("udtestdev-my-new-tls.wallet");
+        assertEquals("0x1586d090e1b5781399f988e4b4f5639f4c2775ef5ec093d1279bb95b9bceb1a0", hash);
+
         hash = resolution.getNamehash("zil");
         assertEquals("0x9915d0456b878862e822e2361da37232f626a2e47505c8795134a95d36138ed3", hash);
         hash = resolution.getNamehash("johnnyjumper.zil");
@@ -112,28 +174,30 @@ public class ResolutionTest {
     }
 
     @Test
-    public void wrongDomainNamehash() throws Exception {
-        TestUtils.expectError(() -> resolution.getNamehash("unupported"), NSExceptionCode.UnsupportedDomain);
-    }
-
-    @Test
     public void getRecord() throws Exception {
-        String recordValue = resolution.getRecord("ryan.crypto", "crypto.ETH.address");
+        String recordValue = resolution.getRecord("testing.crypto", "crypto.ETH.address");
         assertEquals("0x58cA45E932a88b2E7D0130712B3AA9fB7c5781e2", recordValue);
+
+        recordValue = resolution.getRecord("udtestdev-my-new-tls.wallet", "crypto.BTC.address");
+        assertEquals("bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh", recordValue);
+
     }
 
     @Test
     public void noRecord() throws Exception {
-        TestUtils.expectError(() -> resolution.getRecord("ryan.crypto", "invalid.record.value"), NSExceptionCode.RecordNotFound);
+        TestUtils.expectError(() -> resolution.getRecord("testing.crypto", "invalid.record.value"), NSExceptionCode.RecordNotFound);
     }
 
     @Test
-    public void addr() throws Exception {
-        String addr = resolution.getAddress("homecakes.crypto", "eth");
-        assertEquals("0xe7474D07fD2FA286e7e0aa23cd107F8379085037", addr);
+    public void getAddress() throws Exception {
+        String addr = resolution.getAddress("udtestdev--awefawef.crypto", "eth");
+        assertEquals("0x58cA45E932a88b2E7D0130712B3AA9fB7c5781e2", addr);
 
-        addr = resolution.getAddress("brad.crypto", "eth");
-        assertEquals("0x8aaD44321A86b170879d7A244c1e8d360c99DdA8", addr, "brad.crypto --> eth");
+        addr = resolution.getAddress("udtestdev-my-new-tls.wallet", "eth");
+        assertEquals("0x6EC0DEeD30605Bcd19342f3c30201DB263291589", addr, "udtestdev-my-new-tls.wallet --> eth");
+
+        addr = resolution.getAddress("testing.crypto", "eth");
+        assertEquals("0x58cA45E932a88b2E7D0130712B3AA9fB7c5781e2", addr, "testing.crypto --> eth");
 
         addr = resolution.getAddress("johnnyjumper.zil", "eth");
         assertEquals("0xe7474D07fD2FA286e7e0aa23cd107F8379085037", addr, "johnnyjumper.zil --> eth");
@@ -142,87 +206,70 @@ public class ResolutionTest {
     @Test
     public void wrongDomainAddr() throws Exception {
         TestUtils.expectError(() -> resolution.getAddress("unregistered.crypto", "eth"), NSExceptionCode.UnregisteredDomain);
+        TestUtils.expectError(() -> resolution.getAddress("unregistered.wallet", "eth"), NSExceptionCode.UnregisteredDomain);
+        TestUtils.expectError(() -> resolution.getAddress("unregistered.blockchain", "eth"), NSExceptionCode.UnregisteredDomain);
         TestUtils.expectError(() -> resolution.getAddress("unregistered26572654326523456.zil", "eth"), NSExceptionCode.UnregisteredDomain);
     }
 
     @Test
     public void UnknownCurrency() throws Exception {
-        TestUtils.expectError(() -> resolution.getAddress("brad.crypto", "unknown"), NSExceptionCode.UnknownCurrency);
+        TestUtils.expectError(() -> resolution.getAddress("udtestdev-my-new-tls.wallet", "unknown"), NSExceptionCode.UnknownCurrency);
+        TestUtils.expectError(() -> resolution.getAddress("testing.crypto", "unknown"), NSExceptionCode.UnknownCurrency);
         TestUtils.expectError(() -> resolution.getAddress("johnnyjumper.zil", "unknown"), NSExceptionCode.UnknownCurrency);
-        TestUtils.expectError(() -> resolution.getAddress("brad.crypto", "dodge"), NSExceptionCode.UnknownCurrency);
+        TestUtils.expectError(() -> resolution.getAddress("testing.crypto", "dodge"), NSExceptionCode.UnknownCurrency);
         TestUtils.expectError(() -> resolution.getAddress("johnnyjumper.zil", "dodge"), NSExceptionCode.UnknownCurrency);
     }
 
     @Test
-    public void ipfsHash() throws NamingServiceException {
-        String ipfs = resolution.getIpfsHash("brad.crypto");
-        assertEquals("QmdyBw5oTgCtTLQ18PbDvPL8iaLoEPhSyzD91q9XmgmAjb", ipfs);
+    public void getIpfsHash() throws NamingServiceException {
+        String ipfs = resolution.getIpfsHash("testing.crypto");
+        assertEquals("QmRi3PBpUGFnYrCKUoWhntRLfA9PeRhepfFu4Lz21mGd3X", ipfs);
 
         ipfs = resolution.getIpfsHash("johnnyjumper.zil");
         assertEquals("QmQ38zzQHVfqMoLWq2VeiMLHHYki9XktzXxLYTWXt8cydu", ipfs);
-
-        ipfs = resolution.getIpfsHash("reseller-test-udtesting-341567718146.crypto");
-        assertEquals("QmVJ26hBrwwNAPVmLavEFXDUunNDXeFSeMPmHuPxKe6dJv", ipfs);
     }
 
     @Test
-    public void emailTest() throws NamingServiceException {
-        String email = resolution.getEmail("johnnyjumper.zil");
-        assertEquals("jeyhunt@gmail.com", email);
+    public void getEmailTest() throws NamingServiceException {
+        String email = resolution.getEmail("testing.crypto");
+        assertEquals("testing@example.com", email);
     }
 
     @Test
-    public void ownerTest() throws NamingServiceException {
-        String owner = resolution.getOwner("brad.crypto");
-        assertEquals("0x8aad44321a86b170879d7a244c1e8d360c99dda8", owner);
+    public void getOwnerTest() throws NamingServiceException {
+        String owner = resolution.getOwner("testing.crypto");
+        assertEquals("0x58ca45e932a88b2e7d0130712b3aa9fb7c5781e2", owner);
+
+        owner = resolution.getOwner("udtestdev-my-new-tls.wallet");
+        assertEquals("0x6ec0deed30605bcd19342f3c30201db263291589", owner);
 
         owner = resolution.getOwner("johnnyjumper.zil");
         assertEquals("0xcea21f5a6afc11b3a4ef82e986d63b8b050b6910", owner);
     }
 
     @Test
-    public void usdtTest() throws Exception {
-        String domain = "udtestdev-usdt.crypto";
-        String erc20 = resolution.getUsdt(domain, TickerVersion.ERC20);
-        assertEquals("0xe7474D07fD2FA286e7e0aa23cd107F8379085037", erc20);
-        String tron = resolution.getUsdt(domain, TickerVersion.TRON);
-        assertEquals("TNemhXhpX7MwzZJa3oXvfCjo5pEeXrfN2h", tron);
-        String omni = resolution.getUsdt(domain, TickerVersion.OMNI);
-        assertEquals("19o6LvAdCPkjLi83VsjrCsmvQZUirT4KXJ", omni);
-        String eos = resolution.getUsdt(domain, TickerVersion.EOS);
-        assertEquals("letsminesome", eos);
-
-        TestUtils.expectError(
-            () -> resolution.getUsdt("unregistered.crypto", TickerVersion.ERC20),
-            NSExceptionCode.UnregisteredDomain
-        );
-        
-        TestUtils.expectError(
-            () -> resolution.getUsdt("homecakes.crypto", TickerVersion.TRON),
-            NSExceptionCode.RecordNotFound
-        );
-    }
-
-    @Test
-    public void ownerFailTest() throws Exception {
+    public void getOwnerFailTest() throws Exception {
         TestUtils.expectError(() -> resolution.getOwner("unregistered.crypto"), NSExceptionCode.UnregisteredDomain);
+        TestUtils.expectError(() -> resolution.getOwner("unregistered.wallet"), NSExceptionCode.UnregisteredDomain);
     }
 
     @Test
     public void noIpfsHash() throws Exception {
         TestUtils.expectError(() -> resolution.getIpfsHash("unregstered.crypto"), NSExceptionCode.UnregisteredDomain);
-        TestUtils.expectError(() -> resolution.getIpfsHash("pickleberrypop.crypto"), NSExceptionCode.RecordNotFound);
+        TestUtils.expectError(() -> resolution.getIpfsHash("udtestdev-my-new-tls.wallet"), NSExceptionCode.RecordNotFound);
+        TestUtils.expectError(() -> resolution.getIpfsHash("udtestdev--awefawef.crypto"), NSExceptionCode.RecordNotFound);
 
     }
 
     @Test
     public void noEmailRecord() throws Exception {
         TestUtils.expectError(() -> resolution.getEmail("brad.crypto"), NSExceptionCode.RecordNotFound);
+        TestUtils.expectError(() -> resolution.getEmail("udtestdev-my-new-tls.wallet"), NSExceptionCode.RecordNotFound);
     }
 
     @Test
     public void dnsRecords() throws Exception {
-        String domain = "udtestdev-reseller-test-udtesting-875948372642.crypto";
+        String domain = "testing.crypto";
         List<DnsRecordsType> types = Arrays.asList(DnsRecordsType.A, DnsRecordsType.AAAA);
         List<DnsRecord> dnsRecords = resolution.getDns(domain, types);
         assertEquals(2, dnsRecords.size());
@@ -248,11 +295,11 @@ public class ResolutionTest {
             public JsonObject request(String url, JsonObject body) throws IOException {
                 throw cause;
             }
-            
+
         };
         Resolution resolutionWithProvider = Resolution.builder().provider(provider).build();
         TestUtils.expectError(
-            () -> resolutionWithProvider.getAddress("brad.crypto", "eth"),
+            () -> resolutionWithProvider.getAddress("udtestdev-my-new-tls.wallet", "eth"),
             NSExceptionCode.BlockchainIsDown,
             cause
         );
@@ -271,7 +318,7 @@ public class ResolutionTest {
                         object.get("data").getAsString()
                             .equals("0x91015f6b0000000000000000000000000000000000000000000000000000000000000040756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc900000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000001263727970746f2e4554482e616464726573730000000000000000000000000000")
                     ) {
-                        JsonObject answer = new JsonObject();                        
+                        JsonObject answer = new JsonObject();
                         answer.addProperty("jsonrpc", "2.0");
                         answer.addProperty("id", "1");
                         answer.addProperty("method", "eth_call");
@@ -306,15 +353,66 @@ public class ResolutionTest {
 
     @Test
     public void testGetMultiChainAddress() throws Exception {
-        String domainWithMultiChainRecords = "udtestdev-usdt.crypto";
+        String domainWithMultiChainRecords = "testing.crypto";
 
         String erc20 = resolution.getMultiChainAddress(domainWithMultiChainRecords, "usdt", "erc20");
-        assertEquals("0xe7474D07fD2FA286e7e0aa23cd107F8379085037", erc20);
+        assertEquals("0x58cA45E932a88b2E7D0130712B3AA9fB7c5781e2", erc20);
         String tron = resolution.getMultiChainAddress(domainWithMultiChainRecords, "usdt", "tron");
-        assertEquals("TNemhXhpX7MwzZJa3oXvfCjo5pEeXrfN2h", tron);
+        assertEquals("TRMJfXXbmwb3WFSRKbeRgKsYoD8o1a9xxV", tron);
         String omni = resolution.getMultiChainAddress(domainWithMultiChainRecords, "usdt", "omni");
-        assertEquals("19o6LvAdCPkjLi83VsjrCsmvQZUirT4KXJ", omni);
+        assertEquals("1KvzMF2Vjy14d6JGY7dG7vjT5kfpmzSQXM", omni);
+        String eos = resolution.getMultiChainAddress(domainWithMultiChainRecords, "usdt", "eos");
+        assertEquals("karaarishmen", eos);
 
-        
+
+    }
+
+    @Test
+    public void testTokenURIUNS() throws Exception {
+        String tokenUri = resolution.getTokenURI("brad.crypto");
+        assertEquals("https://staging-dot-dot-crypto-metadata.appspot.com/metadata/brad.crypto", tokenUri);
+
+        TestUtils.expectError(() -> resolution.getTokenURI("fake-domain-that-does-not-exist.crypto"), NSExceptionCode.UnregisteredDomain);
+    }
+
+    @Test
+    public void testTokenURIZNS() throws Exception {
+        String testDomain = "brad.zil";
+        TestUtils.expectError(() -> resolution.getTokenURI(testDomain), NSExceptionCode.NotImplemented);
+    }
+
+    @Test
+    public void testTokenURIMetadata() throws Exception {
+        String testDomain = "brad.crypto";
+
+        TokenUriMetadata metadata = resolution.getTokenURIMetadata(testDomain);
+        assertNotNull(metadata);
+        assertEquals(metadata.getName(), testDomain);
+        assertEquals(metadata.getAttributes().size(), 8);
+        TokenUriMetadataAttribute attribute = metadata.new TokenUriMetadataAttribute();
+        attribute.setTraitType("ETH");
+        attribute.setValue("0x8aaD44321A86b170879d7A244c1e8d360c99DdA8");
+        assertTrue(metadata.getAttributes().contains(attribute));
+    }
+
+    @Test
+    public void testUnhashCNS() throws Exception {
+        String testHash = "0x756e4e998dbffd803c21d23b06cd855cdc7a4b57706c95964a37e24b47c10fc9";
+        String tokenName = resolution.unhash(testHash, NamingServiceType.UNS);
+        assertEquals("brad.crypto", tokenName);
+
+    }
+
+    @Test
+    public void testUnhashUNS() throws Exception {
+        String testHash = "0x1586d090e1b5781399f988e4b4f5639f4c2775ef5ec093d1279bb95b9bceb1a0";
+        String tokenName = resolution.unhash(testHash, NamingServiceType.UNS);
+        assertEquals("udtestdev-my-new-tls.wallet", tokenName);
+    }
+
+    @Test
+    public void testUnhashZNS() throws Exception {
+        String testHash = "0x5fc604da00f502da70bfbc618088c0ce468ec9d18d05540935ae4118e8f50787";
+        TestUtils.expectError(() -> resolution.unhash(testHash, NamingServiceType.ZNS), NSExceptionCode.NotImplemented);
     }
 }